Spring-13 updating Flyway

+++ Spring-13-ORM-CinemaLab/src/main/java/com/mryzhan/entity/Location.java
@@ -3,12 +3,14 @@ package com.mryzhan.entity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.ManyToOne;
 
+import java.math.BigDecimal;
+
 @Entity
 public class Location extends BaseEntity {
 
     private String name;
-    private Long latitude;
-    private Long longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
     private String country;
     private String city;
     private String state;
